addVisit 수정하기
- 별점 변수를 int 타입에서 double 타입으로 수정하기

Related to: #37

+++ src/main/java/com/example/naverrestaurant/db/H2DbRepositoryAbstract.java
@@ -111,7 +111,7 @@ abstract public class H2DbRepositoryAbstract<T extends MemoryDbEntity> implement
             if(rs.getTimestamp("lastVisitDate") != null){   // lastVisitDate가 null이 아닐 때만 set 해주기
                 wishListEntity.setLastVisitDate(rs.getTimestamp("lastVisitDate").toLocalDateTime());
             }
-            wishListEntity.setStarRating((rs.getInt("starRating")));
+            wishListEntity.setStarRating((rs.getDouble("starRating")));
             return wishListEntity;
         });
     }
